[pangofc] Be more tolerant on the "fontmap" property setting
Bug 578336 – ABI breakage from 1.22 to 1.24 in fontmap field of PangoFcFont

Commit 9305b9a9995e84ace3818a90346820ef8c47a3ce introduced a "fontmap"
property in PangoFcFont and expected all subclasses to set it upon
construction.  And PangoFcFontSet was changed to not set fcfont->fontmap
anymore.

Relax this change to make it backward compatible by:

  1) Making "fontmap" property a not-construct-only property

  2) In PangoFcFontSet, set "fontmap" if not already set

This should make it backward compatible, while allowing new code to set the
property instead of relying on PangoFcFontSet to do it.