Re-enable introspection, delete PangoX
This patch relies on the latest GObject Introspection; bump
requirement to 0.6.4.

First, delete the PangoX.gir since it's "incredibly deprecated"
in Owen's words =)

Next, clean up the PangoCairo.gir logic; we need to include the .gir
for the relevant font backend (currently just FT2), because the
font object gets expanded.