Revert usage of CoreText in ATSUI backend
Next to CGFontRef we also store an ATSFontRef now.  This required
the addition of the public for backends function
pango_atsui_font_get_atsfont().  I have marked this call to be available
starting version 1.27, as Pango is now in the midst of the 1.26 series.

Since we now store the ATSFontRef next to CGFontRef, we can continue to
make use of the ATS API in
pango_cairo_atsui_font_create_metrics_for_context().  The CoreText calls
have been removed.  Both Tiger and earlier, and Leopard and later machines
will now use the ATS API.

It is well known that the ATS API has been deprecated in Leopard.  In the
near future a brand-new CoreText backend has to be written (from scratch)
that should be installed and used on Macs running Leopard and later.

This patch has been tested on 32-bit Tiger and 64-bit Snow Leopard and
found to be functioning correctly.  The build should now work again fine
on Tiger machines, as it has been broken since the introduction of the
CoreText calls.