Start implementation of CoreText backend
On Mac OS X systems where CoreText is available (that is, 10.5 or
higher), we will build the CoreText backend instead of the ATSUI
backend.  In this case, the ATSUI backend will not be built and
installed.  This change is necessary because since Mac OS 10.5
ATSUI has been marked as deprecated.

For older systems, we will have to continue to use the
ATSUI backend.