Add PANGO_LIBDIR and PANGO_SYSCONFDIR enviroment variables
Add environment variables to override the compile time values
for the libdir and sysconfdir. This provides additional
flexibility and enables using a static pango.modules file
for libraries packaged with an application on OS X by setting
environment variables at application startup and using
paths of the form:

 @executable_path/../lib/pango/modules/..

in the modules file.

https://bugzilla.gnome.org/show_bug.cgi?id=554524