layout: Make sure all lines are ellipsized
All lines that are too long should be ellipsized when ellipsizing is
enabled. Not just those lines where we thought they'd need to be in
advance. It might turn out that a word is too long so that we cannot
wrap it. In that case we still want to break as soon as possible, but
still ellipsize the line.

https://bugzilla.gnome.org/show_bug.cgi?id=649783