CoreText: stricter handling of FontSymbolic traits
It turns out that getting this value as "Int" is incorrect and resulted
in font traits not being returned in some cases.  Without traits, an
italic trait is not set, which caused synthetic oblique fonts to be
created when not necessary.

Also use CTFontSymbolicTraits type in the PangoCoreTextFace structure
and do a stricter bit mask check for certainty.