Bug 664125 - Zero-width spaces cause missing characters
This is fixed by using the characters obtained from CoreText's mapping from
glyphs back to the original string instead of blindly iterating over the
original gchar string. Take notice that we have to convert these indices
from the mapping back to a byte index in the original UTF8 string.

Zero-width spaces are already removed from the list of CGGlyphs, so we were
mismatching CGGlyphs and characters from the original string. Using
CoreText's list of string indices instead eliminates the mismatches.