querymodules: Add --system argument
When run as a system trigger, we really don't want to be consulting
e.g. /root/.pangorc.  Also, the OSTree build system can run as
non-root, but we don't in this case want to look at /home/user since
the user may not even exist in the password database inside the
chroot.

https://bugzilla.gnome.org/show_bug.cgi?id=667960