Bug 673497 - corefont fallback not always working
The fallback failed when a "small caps" font was requested. This commit
improves the fallback support. When the first fallback, trying Sans with
the same style fails, we reset the variant, weight and stretch to
default values and try again. With Sans we should always be able to
adhere to the requested style.

Last but not least, output a sensible error message if all fallbacks
fail instead of simply crashing on a NULL pointer somewhere.