Rework config file reading
The code was loading and parsing the system config file
every single time, and has not correctly handling mixed
requests for user and system config entries.

This commit reworks the code so that the configuration is
loaded only once, in threadsafe fashion.
pango_config_key_get_system is no longer using the same
hash table, but reloads its data every time - this is
not a really problem, since this function is only used
in pango-querymodules.