Prepare Visual Studio 2010 projects for Visual Studio 2012
This adds a PlatformToolset tag in the project configs so that we can
provide support for Visual Studio 2012 with relative ease as the format
of the VS 2012 projects are only slightly different from their VS 2010
counterparts. We can then use a script like the one used in GLib[1] to
copy the VS2010 projects and replace the necessary tags to create the
VS2012 projects.

This also cleans up the projects and property sheets, as there were
some unwanted/unneeded entries in them.

[1]: http://git.gnome.org/browse/glib/commit/?id=76cecf061b377d30e5422cdddb1fb9d19c52421d