Revert "Fix Win32 and CoreText cairo backends to take cairo CTM into account"
This reverts commit b359bb484bc7d27a835cfd6d9ffcaed582ebf2ae.

Bug 694918 - Pango layout is rendered in wrong size if we have
non-identity cairo transformation matrix

I was wrong blindly copying this logic from pangocairo-fc into
pangocairo-win32 and pangocairo-quartz.

The reason we need the it in pangofc is because the fontsize we
get back from fontconfig is multiplied by the ctm scale factor,
so we undo it when loading the font.  The same is NOT true about
the non-fc backends, so the copying was totally wrong.