Bug 707131 - removing useless test.
Since pango_scan_int() would return FALSE if the scanned value was over
INT_MAX; and the reference out value's type is an int, (n > INT_MAX)
test is unecessary.
If pango_scan_int() behavior was to be changed to accept long
int, an explicit test should be added (or the PangoAttrSize
type updated to handle long int too).