MSVC Builds: Rework Introspection Build
The current approach of building the introspection files for Pango works,
but is often cumbersome as one needs to set many environmental variables
before launching a solution file, which runs a Windows batch script to
generate the .gir/.typelib files. It was also possible to hand-run the
batch script from the Visual Studio command prompt, but even more
environmental variables need to be set.

This changes the approach to build the introspection files using an NMake
Makefile (but elimating from the Visual Studio Project Files the part to
build the introspection files) to:

-Make it clearer to the person building the introspection files what
 environmental variables are needed, specifically for PKG_CONFIG_PATH and
 MINGWDIR and CFG (formerly CONF). Setting stuff like VSVER, PLAT and
 BASEDIR is no longer required, which was a bit clunky.
-Allows some more easier flexibility on the build of the intropsection
 files.