MSVC Introspection Build: Fix build
Link and run the dumper program firstly against the build of ATK from
the same tarball, then linking to the other dependent libraries which
this build of ATK was built against before finally trying to link and
run against other existing builds.

This is especially important for unstable releases, as APIs are often
added and dependent upon during development cycles.