Build: Provide Simple Support for MSVC 2012/2013
As the Visual Studio 2012/2013 are only slightly different from the Visual
Studio 2010 projects, we can provide support for them by using scripts to
copy the Visual Studio 2010 projects, and update the specific parts as
necessary.  Thus, there would be little maintenance overhead for these as
only the 2010 projects need to be kept up-to-date as a result.

This might change when we do get the stack working with WinRT/Metro, but
that's going to be another totally different issue.