Add tests for pango markup
This test is modeled after the GMarkup parser tests in GLib.
Test file with a name ending in .markup can be dumped in
tests/markups/, and tests/markup-parse can be used to generate
the expected output files which should be in the same location
and with the extension .markup replaced by .expected.

The test is also used as an installed test.