build/Makefile.msvcproj: Don't Use realpath(1)
Apaprently there are many incompatible realpath(1) implementations out
there, so things like --relative-to can't be assured to work everywhere,
such as on Ubuntu.

Instead, simply use the subst feature of automake to fulfil the same
purpose, which is like what the rest of this file does, and follows the
pattern of the introspection makefile, which this is modelled on.

This is also checked with 'make -j8 distcheck'

https://bugzilla.gnome.org/show_bug.cgi?id=735039