README.win32: Update Information for Builds
Since module functions are now built into the main Pango DLLs,
indicate so in README.win32, although it does not make any difference
in the Visual Studio builds on the surface, as modules are already
built into the main DLLs for such builds.

Also update the information on the Visual Studio versions supported
by the project files.