MSVC Builds: Use Autotools Module for Introspection
This uses the newly-added autotools module so that the file list
for introspection using Visual Studio is now generated during
'make dist', along with the g-ir-scanner command line to generate
the .gir files and the compilation of the .typelib files.

This would help to simplify maintenance of the introspection
builds in Visual Studio, which is done using NMake.

https://bugzilla.gnome.org/show_bug.cgi?id=764984