Export symbols with compiler directives if possible
Add a configure check to see whether compiler directives are available
for exporting symbols, and use them if so.  Likewise, update
the Visual Studio projects and config.h.win32.in to do likewise for
Windows builds.

We can then drop the .def files that were used to export symbols on
Windows builds, which should clean up things a bit.