Visual Studio builds: Move projects to win32/
This allows one to go down one less level in the directory tree to get to the
Visual Studio project files, and so make things more in line with the rest of
the GTK+ stack.  This also cleans up the Visual Studio 201x projects as there
are some items that can be actually combined.