Visual Studio builds: Support Visual Studio 2017
This updates the autotools scripts to support Visual Studio 2017 builds, by copying
the Visual Studio 2010 projects and updating items in there as necessary.

Note the format of the toolset version changed, so we allow one to specify a
custom toolset version string and hence use it if it is specified, otherwise
the toolset version string is generated as it was before.

Since Visual Studio 2017 aims to be compatible on the CRT level as Visual Studio
2015, the 2017-compiled binaries should be usable without problems with 2015-compiled
binaries.