pango: Remove some unnecessary NULL checks
These checks are preceded by other checks or code which demonstrates
that the values are definitely always non-NULL.

In the case of pango_attr_list_insert_internal(), the final branch of
the function never needs to update list->attributes_tail, as the middle
branch handles the case of appending to the list. The final case is just
for insertions in the middle of the list.

Coverity IDs: 1391710, 1391711

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://bugzilla.gnome.org/show_bug.cgi?id=778654