Correctly implement version macros for Pango
First of all, we need to define `PANGO_UNAVAILABLE`, as it's missing a
definition, and the only reason why nobody caught this is because we're
building without -Wundef, and clearly nobody attempted using
min-required/max-allowed with Pango.

Additionally, we really need a way to disable deprecation warnings from
the compiler.

The complete implementation of the version macros for GLib-based
libraries includes both items above.