meson: Check for HarfBuzz and FontConfig for PangoFT2
It is possible that we can have the following situations, at least on
Windows:
-FreeType present, FontConfig missing
-Cairo-FT present, with no FontConfig support.

As gen-script-for-lang requires FontConfig, and PangoFT2 depends on
HarfBuzz, FontConfig and Freetype, we need to check for them before we
build PangoFT2, and so that we could include PangoFT2 support in
PangoCairo.

The tests and pango-view have an optional dependency on PangoFT2, so we
need to also check whether we built PangoFT2 before we try to build
things related to PangoFT2.

For the tools, since gen-script-for-lang.c depends on FontConfig, check
for it as well before we build it.

https://bugzilla.gnome.org/show_bug.cgi?id=783274