break: Stop lumping ZWJ with Extend
ZWJ is treated as its own class in TR29, and us lumping
them together and then manually checking wc == 0x200d in
various places was causing us to inadvertedly inserting
grapheme breaks in the middle of Emoji ZWJ sequences
where they are not suppose to be.

Add test cases to verify this.