build: Integrate introspection build in MSVC projects
This adds the introspection NMake Makefiles buildable from within the
Visual Studio projects, so that the whole Visual Studio build process
would be simplified.  At its current state, this will build the
introspection files for PangoCairo and Pango, but does not include
PangoFT2 support, so this is not supported in the [Release|Debug]_FC
configs.  For this to work, one needs to have a complete installation of
GObject-introspection in $(GlibEtcInstallRoot), and will need to ensure
that PythonDir (or PythonDirX64 on x64 systems) corresponds to the
same Python series that was used to build GObject-Introspection.

This needs to be built (and cleaned) separately as this is not built by
default, by building/cleaning the pango-introspect project directly.

In this commit, also rename the PythonPath entries in the projects and
property sheets, so that we avoid getting confused with the PYTHONPATH
envvar; and for Visual Studio 201x builds, we use Python 3.4.x by
default 2010, 2012 and 2013, and 3.6.x by default for 2015 and 2017, as
what is done now in GObject-Introspection projects.