Avoid a memory leak
We call pango_fc_font_key_init on a stack-allocated temporary
lookup key, with the expectation that it does not have to be
freed. In the case where we want to use it as key in the font_hash,
we call copy() on it. So, don't duplicate the variations string
in init() - it will get duplicated in copy() later.