build: move usp10 before gdi32
Without this pango on mingw64 tries to lookup up the Script* functions
in gdi32 and fails. It already fails at the build stage because the
introspection dump crashes with a missing entry point error.

Moving usp10 before gdi32 makes things work.

This might be related to the warning in the uniscribe docs:
"Important Starting with Windows 8: To maintain the ability to run on
Windows 7, a module that uses Uniscribe must specify Usp10.lib
before gdi32.lib in its library list."

https://bugzilla.gnome.org/show_bug.cgi?id=795045