pangowin32: fix script cache hash key for 64bit builds
It joins the HFONT and script key to a gint64 and uses this as a
hash key, but HFONT is a pointer type and on 64bit Windows it
doesn't fit in 32bit and the value gets truncated.

This breaks the build with meson where -Werror=pointer-to-int-cast
is enabled by default.

Instead of using the gint64 hash functions add our own key type and implement
matching hash and equality functions for it.

https://bugzilla.gnome.org/show_bug.cgi?id=795045