Short-circuit FriBidi call if paragraph is unidirectional
This was included in Pango's mini-fribidi but removed when we moved to
external FriBidi.

Most apps create PangoLayout twice to show text, one to measure, one to
render.  Each PangoLayout shaping apparently calls FriBidi twice (TODO:
figure out why and fix); FriBidi creates two runs per work.  So that's
eight malloc calls per word to show text.  That's a lot.  With this
optimization that completely disappears for most text.

We should make an API in FriBidi out of this.

Reported by Christian Hergert.