build: Remove Visual Studio projects
We have been able to build Pango with Visual Studio using Meson for some
time now, so drop the Visual Studio projects as well.  This should also
make builds more flexible as the Visual Studio projects focused more on
the cases where we do not build PangoFT2, especially in regards to
building the introspection files.