WIP: Replace PangoScript by GUnicodeScript
This commit replaces the PangoScript enum by a
typedef to GUnicodeScript, and preserves the old
enum values as defines. It can't quite be merged
as is because we rely in some places on being able
to get string names for enum values, and for that
we need to register GUnicodeScript as an enum
type, in GObject.