pango: th_brk_new() may return NULL.
The call to th_brk_new() may return NULL, in case of allocation failure
of course, but also if the default dictionnary fails to load, which may
be an unlikely event, yet far from impossible (packaging may have
forgotten the file hence it may be absent, have been corrupted, or
whatever else).
In any case, although unlikely, it is not impossible, and should not
crash the whole program so we should check the return value and output a
warning if NULL.

Note that the call to th_brk_find_breaks() can still run even if brk is
NULL so we only check before freeing the data.