fc: Move font setup code from the shaper
Move all the code that sets up the hb_font_t to
pango_fc_font_create_hb_font, and use it from
the shaper. This is the second step towards
taking over font management.

Even better, harfbuzz has a ready-made function
for this. We can drop a lot of FT_Face-using code
this way. We assume unhinted rendering for now,
so we can set ppem to 0.