Deal with Common scripts as und-zyyy language tag.
When fonts being rendered for Common has different looks to Latin,
Pango changes a font and makes different looks for a cluster of words
just after tying a character. e.g. 012345 vs 012345a.

This is a long standing UX problem in Pango and they should be rendered
with the same font all the time.

This change tries to fix this issue and leave customization to fontconfig
layer.

Fixes https://gitlab.gnome.org/GNOME/pango/issues/353