Core text: Fix indexing into cascade list when multiple fonts are selected
Bug introduced from 633fd80.

With that changeset, if multiple fonts from the provided font family
are found, they are added to the start of fontset->fonts. Space for
the cascade list (based on the 'best'/first font) then follows.

However, pango_core_text_fontset_get_font_at was not updated to
match this change. So it still assumes that there is at most only
one 'real' font in fonset->fonts. With this assumption, the
indexing into fontset->cascade_list became incorrect, leading to
an out-of-bounds access on that list, resulting in an exception
being thrown.

Track the number of fonts found in pango_core_text_fontset_new
in real_fount_count, and use this as the offset into cascade_list.

This also fixes a memory leak; when a PangoCoreTextFontSet is
initialised, fonts and coverages are pre-initialised, but this
was being discarded and re-set in pango_core_text_fontset_new
without freeing the previous initialisation.