layout: Don't zero the width of visible space
Don't zero out a line separator at the end of line
if we turned it into an unknown glyph to render
it.

And if we zero out whitespace at the end of the line,
make sure we don't draw anything there by setting
the glyph to EMPTY.

Without this, we are getting the [LS] hex box
rendered on top of the last character in a line,
when the line separator ends up being visible.