itemize: Don't change font for line separator
If we render this (in single-paragraph mode), we
draw a hex box, so changing font just for the line
separator is never useful, and in fact, counterproductive
in cases where the font has a (typically empty) glyph
for it.

This fixes the [LS] hex box not showing up in
single-paragraph mode depending on font fallback.