itemize: Don't break runs for font_features
We don't need to break runs just because font_features
changed - we pass those to harfbuzz with the range they
apply to, after all.

To do this, split the attributes into two lists,
those that break runs (as before), and those that
don't. And collect the overlapping attributes from
the non-breaking list at the end of processing an
item, when we know the range.

For now, only font_features is considered non-breaking,
but this can be expanded.

Closes: https://gitlab.gnome.org/GNOME/pango/issues/242