layout: Fix some offset-vs-index confusion
We had some sad cases where we used a char offset
as a byte index, with the expected bad results once
there are multibyte characters in the mix.

Instead of poking at the text in the middle of
line-breaking, go back to keeping a plain array
of positions that need a hyphen inserted. We
maintain this array in parallel to the existing
log_widths array, so hopefully the offset math
is working out now.

It is tempting to look at log_attrs.is_white for this,
but that gives us double hyphens when breaks are taken
at explicit hyphens. So, keep this information separately.