build: Only use fallback dependency for HarfBuzz when needed
Unfortunately the CMake dependency discovery mechanism is broken in
Meson, so on MSVC builds we still look first for the HarfBuzz headers
and .lib's, and ensure that they are 2.0.0 or later, before we try to
use the fallback dependency.

We do, however, want to make use of the fallback if HarfBuzz cannot be
found since HarfBuzz is now a hard dependency of Pango for all builds.