PangoWin32: Use HarfBuzz to get font coverage
The Windows code was manually reading the font cmap table, so we may as
well just use HarfBuzz here since it has better cmap support.

I also removed the code that did some hacks to check for font name
entries in certain languages and drop some characters from the cover
based on that. This looked very suspicious and was added in 2002 with no
much explanation. We don’t do anything like this on Linux or macOS, so
better be consistent here.