Take font matrix into account
When setting up the scale of the hb_font_t, we need
to take both the pango ctm and the font matrix into
account - this used to come for free when we were
calling into cairo for getting glyph metrics. Now,
we use harfbuzz for glyph metrics, so we need to
give it the proper scale info.

The symptom of this was Emoji getting an enormous
width. Curiously, cairo would still render them
at the expected size.