coretext: Make “system-ui” alias work
Alias names need to be inserted into the font map or otherwise will be
ignored and mapped to Sans (in pango_core_text_font_map_load_fontset()).

Furthermore, there is no font named ”San Francisco” on macOS, the actual
UI font is a hidden font called ”.SF NS Text” or something like that,
and there an “.AppleSystemUIFont” alias that should be used instead.