Pass the right attributes when shaping ellipses
Now that we are splitting attributes into those that
are relevant for itemization and shaping, we need to
make sure to pass the right ones along when ellipsizing,
or we risk picking a wildly mismatching font for the
ellipsis run, causing things to shift vertically.

Test included.

Closes: https://gitlab.gnome.org/GNOME/pango/issues/397

Thanks to Jorge Luis Martinez Gomez for his help
in tracking this down.