PangoLanguage: Fix clang -Wcast-align warnings
Clang complains when casting char* to PangoLanguage* and -Wcast-align
set. We only cast it to char* to get the proper address for the private
struct though, so just cast to void* afterwards.

Except for the case that was stuffing PangoLanguage* sized elements into
a GArray. Just use a GPtrArray there.