Move FcFontMatch and FcFontSort calls to a thread
fontconfig is thread-safe, so we can do calls that
can take several milliseconds to complete in a thread.
This patch does that for FcFontSetMatch and FcFontSetSort.
It is a win, at least for FcFontSetSort, since it
reduces the time that the main thread spends in
pango_fc_patterns_get_font_pattern for i > 0 from
around 10 to 6 milliseconds.