Fix attr iterators with overlapping attributes
This broke when PangoAttrIterator was changed to
use an array for the stack of current attributes, in
dec6c0868ef2c36. We were not always walking the array
in the right order. The first attribute to check is
at the *end* of the array.

This showed up as misrendering in epsilon_0 example
in gtk3-demo's text view markup demo.

Test included.