build: Remove deprecated python3 Meson module use
We are using the deprecated python3 module for historical reasons. We
used to support both Python 2 and Python 3 in the Autotools build, but
since we switched to Meson we *know* we have Python 3 available. This
allows us to use a shebang line in the Python scripts we use in the
build, instead of invoking them through the Python interpreter in a
custom target; Meson knows how to handle shebangs portably as well.

This change removes the only deprecation warning coming from Meson when
configuring the Pango build.