ft2, xft: Write out the full type of the SubstituteFunc
If the SubstituteFunc used for the now-deprecated set_default_substitute
functions is an alias for PangoFcSubstituteFunc, GObject-Introspection
doesn't realise it's a function pointer, and doesn't flag the user data
and destroy-notify arguments as such. This results in bindings like
PyGI thinking that they are entirely separate arguments, which is an
introspection API break (and probably not something that can
practically be called any more).

Signed-off-by: Simon McVittie <smcv@debian.org>
Resolves: https://gitlab.gnome.org/GNOME/pango/-/issues/510