fc: Fix an ordering issue
With the recently introduced threading for fontconfig
calls, we introduced an ordering issue where the fontmap
may die before an outstanding thread returns, and then
the code that unrefs the pattern object tries to remove
it from the fontmap cache.

Prevent that by giving each thread a strong ref on the
fontmap while it runs.

Fixes: #537