build: Don't build HarfBuzz prematurely
For Visual Studio builds, since HarfBuzz is built with CMake until
HarfBuzz 2.6.x yet we only require HarfBuzz 2.0.0 or later, we
aren't able to find pre-2.6.x HarfBuzz using Meson's dependency()
method as the support for CMake config files is not done in Meson.

As a result, we need to make sure that we first manually look for the
HarfBuzz/HarfBuzz-GObject libraries before we attempt to use the
fallback mechanism in Meson, so that we won't rebuild HarfBuzz
unneccesarily.

This rebuild will also cause issues for older compilers that don't
have enough support for C++11 as building Harfbuzz with Meson support
requires a C++11 compiler, which was mandated in HarfBuzz 2.5.0.